+++ projects/broccolium-fabric/src/test/kotlin/site/siredvin/broccolium/test/FabricFluidStorageTests.kt
@@ -21,7 +21,7 @@ internal class FabricFluidStorageTests : FluidStorageTests() {
                         SingleFluidStorage.withFixedCapacity(1000L * PlatformToolkit.get().fluidCompactDivider.toLong()) {}
                     if (!stack.isEmpty) {
                         Transaction.openOuter().use {
-                            storage.insert(stack.toVariant(), stack.platformAmount, it)
+                            storage.insert(stack.toVariant(), stack.platformAmount.toLong(), it)
                             it.commit()
                         }
                     }
@@ -46,7 +46,7 @@ internal class FabricDummyFluidStorageTests : FluidStorageTests() {
                         SingleFluidStorage.withFixedCapacity(1000L * PlatformToolkit.get().fluidCompactDivider.toLong()) {}
                     if (!stack.isEmpty) {
                         Transaction.openOuter().use {
-                            storage.insert(stack.toVariant(), stack.platformAmount, it)
+                            storage.insert(stack.toVariant(), stack.platformAmount.toLong(), it)
                             it.commit()
                         }
                     }
@@ -71,7 +71,7 @@ internal class FabricReverseDummyFluidStorageTests : FluidStorageTests() {
                         SingleFluidStorage.withFixedCapacity(1000L * PlatformToolkit.get().fluidCompactDivider.toLong()) {}
                     if (!stack.isEmpty) {
                         Transaction.openOuter().use {
-                            storage.insert(stack.toVariant(), stack.platformAmount, it)
+                            storage.insert(stack.toVariant(), stack.platformAmount.toLong(), it)
                             it.commit()
                         }
                     }
