+++ projects/tweakium-core/src/main/kotlin/site/siredvin/tweakium/modules/plugins/FluidStoragePlugin.kt
@@ -3,4 +3,4 @@ package site.siredvin.tweakium.modules.plugins
 import net.minecraft.world.level.Level
 import site.siredvin.broccolium.modules.storage.fluid.api.AgnosticFluidStorage
 
-open class FluidStoragePlugin(level: Level, override val storage: AgnosticFluidStorage, fluidStorageTransferLimit: Int) : AbstractFluidStoragePlugin(level, fluidStorageTransferLimit)
+open class FluidStoragePlugin(level: Level, override val storage: AgnosticFluidStorage, fluidStorageTransferLimit: Double) : AbstractFluidStoragePlugin(level, fluidStorageTransferLimit)
